POLIO-1037: convert RoundForm to tsx
